+++ lib/admanager_web.dart
@@ -1,7 +1,8 @@
 library admanager_web;
 
 /// Simple plugin to display ads from AdManager in Flutter Web
-/// This was primarily created to display Adsense ads on Flutter Web
+/// This was primarily created to display Adsense Ads on Flutter Web,
+/// but it should work with any AdManager Ad.
 export 'src/adblock.dart' show AdBlock;
 export 'src/adblocksize.dart' show AdBlockSize;
 export 'src/admanager.dart' show AdManagerWeb;