+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.2
+
+- Check if the GPT.js Library is already loaded
+- Add VisibilityDetector to the AdManagerBanner to request ads only when the widget is visible
+- Allow to request ads from multiple sizes by block
+- Add EventListener to the GPT.js Library to listen to the `slotRenderEnded` event to hide the ad if it is empty
+
 ## 0.0.1
 
 - Create one simple way to display Ads from AdManager in Flutter Web